backend/integration_module/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
slack-sdk
python-dotenv
pymupdf
python-docx
